Benchmark/Social Network/functions/upload-user-mentions/requirements.txt
pymongo==3.11.0
python-memcached
